requirements.txt
PyYAML==5.4.1
slack-sdk==3.10.1
